Foi criado método de excluir.

+++ src/main/java/br/com/drogaria/drogariaspring/repository/CategoriaRepository.java
@@ -3,5 +3,7 @@ package br.com.drogaria.drogariaspring.repository;
 import br.com.drogaria.drogariaspring.domain.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaRepository extends JpaRepository<Categoria, Byte> {
+import java.util.Optional;
+
+public interface CategoriaRepository extends JpaRepository<Categoria, Short> {
 }
